requirements.txt
fastapi==0.120.4
uvicorn[standard]==0.38.0
sqlmodel==0.0.27
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pytest==8.4.2
httpx==0.26.0
